Safari-611.1.10/LayoutTests/platform/mac/media/audio-session-deactivated-when-suspended-expected.txt


RUN(internals.settings.setShouldDeactivateAudioSession(true))
RUN(video = document.querySelector("video"))
RUN(video.src = findMediaFile("video", "../../../media/content/test"))
RUN(video.play())
EVENT(playing)
EXPECTED (internals.audioSessionActive() == 'true') OK
RUN(internals.processWillSuspend())
EXPECTED (internals.audioSessionActive() == 'false') OK
RUN(internals.processDidResume())
EXPECTED (internals.audioSessionActive() == 'true') OK
END OF TEST
